Move AES-GCM-SIV implementation into rust/crypto
