--- rust/aes-gcm-siv/Cargo.toml
@@ -1,36 +1,0 @@
-#
-# Copyright (C) 2020 Signal Messenger, LLC.
-# SPDX-License-Identifier: AGPL-3.0-only
-#
-
-[package]
-name = "aes-gcm-siv"
-version = "0.1.0"
-authors = ["Jack Lloyd <jack@signal.org>"]
-edition = "2018"
-license = "AGPL-3.0-only"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-aes-soft = "0.6"
-polyval = "0.4"
-subtle = "2.3"
-cipher = "0.2"
-generic-array = "0.14"
-
-[target.'cfg(all(target_arch = "aarch64", any(target_os = "linux")))'.dependencies]
-libc = "0.2" # for getauxval
-
-[target.'cfg(any(target_arch = "x86", target_arch = "x86_64"))'.dependencies]
-aesni = { version = "0.10", features = ["nocheck"] }
-
-[dev-dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-hex = "0.4"
-criterion = "0.3"
-
-[[bench]]
-name = "aes_gcm_siv"
-harness = false
