+++ rust/bridge/ffi/Cargo.toml
@@ -16,7 +16,6 @@ crate-type = ["staticlib"]
 
 [dependencies]
 libsignal-protocol = { path = "../../protocol" }
-aes-gcm-siv = { path = "../../aes-gcm-siv" }
 device-transfer = { path = "../../device-transfer" }
 signal-crypto = { path = "../../crypto" }
 libsignal-bridge = { path = "../shared", features = ["ffi"] }
