+++ rust/bridge/shared/src/ffi/error.rs
@@ -6,7 +6,6 @@
 use std::convert::TryFrom;
 use std::fmt;
 
-use aes_gcm_siv::Error as AesGcmSivError;
 use device_transfer::Error as DeviceTransferError;
 use libsignal_protocol::*;
 use signal_crypto::Error as SignalCryptoError;
@@ -15,7 +14,6 @@ use signal_crypto::Error as SignalCryptoError;
 #[derive(Debug)]
 pub enum SignalFfiError {
     Signal(SignalProtocolError),
-    AesGcmSiv(AesGcmSivError),
     DeviceTransfer(DeviceTransferError),
     SignalCrypto(SignalCryptoError),
     InsufficientOutputSize(usize, usize),
@@ -29,9 +27,6 @@ impl fmt::Display for SignalFfiError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
             SignalFfiError::Signal(s) => write!(f, "{}", s),
-            SignalFfiError::AesGcmSiv(c) => {
-                write!(f, "AES-GCM-SIV operation failed: {}", c)
-            }
             SignalFfiError::DeviceTransfer(c) => {
                 write!(f, "Device transfer operation failed: {}", c)
             }
@@ -56,12 +51,6 @@ impl fmt::Display for SignalFfiError {
 impl From<SignalProtocolError> for SignalFfiError {
     fn from(e: SignalProtocolError) -> SignalFfiError {
         SignalFfiError::Signal(e)
-    }
-}
-
-impl From<AesGcmSivError> for SignalFfiError {
-    fn from(e: AesGcmSivError) -> SignalFfiError {
-        SignalFfiError::AesGcmSiv(e)
     }
 }
