+++ rust/bridge/shared/src/jni/error.rs
@@ -7,7 +7,6 @@ use jni::objects::{GlobalRef, JObject, JString, JThrowable, JValue};
 use jni::{JNIEnv, JavaVM};
 use std::fmt;
 
-use aes_gcm_siv::Error as AesGcmSivError;
 use device_transfer::Error as DeviceTransferError;
 use libsignal_protocol::*;
 use signal_crypto::Error as SignalCryptoError;
@@ -18,7 +17,6 @@ use super::*;
 #[derive(Debug)]
 pub enum SignalJniError {
     Signal(SignalProtocolError),
-    AesGcmSiv(AesGcmSivError),
     DeviceTransfer(DeviceTransferError),
     SignalCrypto(SignalCryptoError),
     Jni(jni::errors::Error),
@@ -33,7 +31,6 @@ impl fmt::Display for SignalJniError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
             SignalJniError::Signal(s) => write!(f, "{}", s),
-            SignalJniError::AesGcmSiv(s) => write!(f, "{}", s),
             SignalJniError::DeviceTransfer(s) => write!(f, "{}", s),
             SignalJniError::SignalCrypto(s) => write!(f, "{}", s),
             SignalJniError::Jni(s) => write!(f, "JNI error {}", s),
@@ -56,12 +53,6 @@ impl fmt::Display for SignalJniError {
 impl From<SignalProtocolError> for SignalJniError {
     fn from(e: SignalProtocolError) -> SignalJniError {
         SignalJniError::Signal(e)
-    }
-}
-
-impl From<AesGcmSivError> for SignalJniError {
-    fn from(e: AesGcmSivError) -> SignalJniError {
-        SignalJniError::AesGcmSiv(e)
     }
 }
