+++ rust/bridge/shared/src/jni/mod.rs
@@ -8,7 +8,6 @@ extern crate jni_crate as jni;
 use jni::objects::{JThrowable, JValue};
 use jni::sys::jobject;
 
-use aes_gcm_siv::Error as AesGcmSivError;
 use device_transfer::Error as DeviceTransferError;
 use libsignal_protocol::*;
 use signal_crypto::Error as SignalCryptoError;
@@ -106,8 +105,6 @@ fn throw_error(env: &JNIEnv, error: SignalJniError) {
         }
 
         SignalJniError::Signal(SignalProtocolError::InvalidArgument(_))
-        | SignalJniError::AesGcmSiv(AesGcmSivError::InvalidInputSize)
-        | SignalJniError::AesGcmSiv(AesGcmSivError::InvalidNonceSize)
         | SignalJniError::SignalCrypto(SignalCryptoError::UnknownAlgorithm(_, _))
         | SignalJniError::SignalCrypto(SignalCryptoError::InvalidInputSize)
         | SignalJniError::SignalCrypto(SignalCryptoError::InvalidNonceSize) => {
@@ -147,7 +144,6 @@ fn throw_error(env: &JNIEnv, error: SignalJniError) {
         | SignalJniError::Signal(SignalProtocolError::SignatureValidationFailed)
         | SignalJniError::Signal(SignalProtocolError::BadKeyType(_))
         | SignalJniError::Signal(SignalProtocolError::BadKeyLength(_, _))
-        | SignalJniError::AesGcmSiv(AesGcmSivError::InvalidKeySize)
         | SignalJniError::SignalCrypto(SignalCryptoError::InvalidKeySize) => {
             "org/whispersystems/libsignal/InvalidKeyException"
         }
@@ -162,7 +158,6 @@ fn throw_error(env: &JNIEnv, error: SignalJniError) {
         | SignalJniError::Signal(SignalProtocolError::InvalidProtobufEncoding)
         | SignalJniError::Signal(SignalProtocolError::ProtobufDecodingError(_))
         | SignalJniError::Signal(SignalProtocolError::InvalidSealedSenderMessage(_))
-        | SignalJniError::AesGcmSiv(AesGcmSivError::InvalidTag)
         | SignalJniError::SignalCrypto(SignalCryptoError::InvalidTag) => {
             "org/whispersystems/libsignal/InvalidMessageException"
         }
